Round 3 Phase 1-3: Backend runtime crashes + HIGH + Infrastructure
Backend Critical:
- B1: notifications.py items=items -> items=item_dicts
- B2: loyalty.py add locale param to get_loyalty_account
- B3: inventory.py data.name -> data.category_name
- B4: reservations.py remove dead code after return
- B5: orders.py fix percentage discount compounding bug
- B6: admin/auth.py guard duplicate RoleAssignment (elif)
- B7: services/order.py select(Order.id) -> func.count(Order.id)

Backend HIGH:
- B9: deps.py hardcoded HS256 -> settings.jwt_algorithm
- B10: staff_ops.py use centralized verify_password (settings argon2)
- B11: staff_ops.py standardize JWT sub to admin.id
- B12: staff_verify_pin uses CurrentStaff not CurrentAdmin
- B13: wallet.py _get_default_currency strips JSONB quotes
- B14: wallet.py admin_topup uses AdminTopupRequest schema
- B15: payment.py wire Stripe/GrabPay webhook HMAC verification
- B16: services/auth.py customer refresh uses TokenBlacklist DB

Infrastructure:
- I11: event_rsvps.customer_id INTEGER -> BIGINT
- I12: GIST indexes cast longitude/latitude to float
- I13/I14: docker-compose healthcheck 3000 -> 3012
- I15: Add updated_at triggers for iam_principals, reservations, staff_shifts
- I16: Seed store 1 operating hours

+++ v3/infra/db/03_tables.sql
@@ -1093,7 +1093,7 @@ CREATE TABLE event_cards (
 CREATE TABLE event_rsvps (
     id SERIAL PRIMARY KEY,
     event_id INTEGER NOT NULL REFERENCES event_cards(id) ON DELETE CASCADE,
-    customer_id INTEGER NOT NULL REFERENCES customers(id) ON DELETE CASCADE,
+    customer_id BIGINT NOT NULL REFERENCES customers(id) ON DELETE CASCADE,
     created_at TIMESTAMPTZ NOT NULL DEFAULT now()
 );
